Move:add-resultをpagesフォルダに移動 #27
