+++ src/app/pages/add-result/add-result.component.scss
